pyproject.toml
[project]
name = "adk-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.11.0",
    "litellm>=1.75.8",
    "openai<1.100.0",
    "ruff>=0.12.9",
    "pymupdf>=1.23.0",
    "rank-bm25>=0.2.2",
]
